Support 'ctx.on.event_name' for specifying events.
Fix typo in comment.

Remove support for directly running custom events.

Update tests and docs to match final (hopefully\!) API decision.

Style fixes.

Fix tests.

The failing test are (a) ones that need to be rewritten for the new system, (b) ones to do with custom events.

Remove the old shortcuts on state components.

Remove old event properties from Secret.

Style fixes.

Move the checks that were on binding to the consistency checker.

Update rubbish event tests.

These are not as valuable without support for emitting custom events directly, but it seems worthwhile to leave them in so that if we add custom event emitting back in the future they're here to build off.

Update tests now that emitting custom events is not possible.

Minor clean-up.

Fix typo found in review.

Fix tests for relation.unit.